Genericise packet handling using GenEvent
We want to build multiple handlers for our packet capture, we may log to
a file or write a pcap file. We may even want to write it to a buffer
later and dump that to a file when called.

By wrapping epcap in with a generic PacketHandler that implements
GenEvent, we provide a simple interface for each implementation to use.
Each implementation just iterates over an infinant stream and handles
each packet